requirements.txt
urllib3
requests
bs4
pycryptodome
.
